requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-nested-routers==0.93.4
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
